requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-axes==5.40.1
django-ipware==4.0.2
django-mathfilters==1.0.0
django-rq==2.4.1
django-translation-manager==1.3.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
nepali-datetime==1.0.7
numpy==1.24.1
panda==0.3.1
pandas==1.5.2
Pillow==9.4.0
polib==1.1.1
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.0
requests==2.28.1
rq==1.11.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
